Secure-inventory-management-store
This project was built in Python to illustrate the usability and flexibility of flask in combination with SQL-Alchemy by designing an Inventory and Sales Management System that handles CRUD operations efficiently and follows REST architectural styles. The system consisted of three components-front end, back end, and flask framework. The backend was developed using a pythonic domain ORM using SQL-Alchemy with the database in SQLite3. The front-end using HTML & CSS was created. Flask provided a smooth link to database and front-end REST architecture. Upon authentication, the client or user can access the website via his credentials. The user can search all the items in the inventory available, quantity, barcode etc. Furthermore, addition, deletion, updating, sale, and search of products can be done as well.